Add implementation of proxy descriptors.
We now have proxy descriptors that are able to be used to load the
messages file blocks into, and will be freed by MessageTrans itself.
These shouldn't be used often, but they're necessary for some modules.

We have support for normalising the output more as well - the error
messages are hidden, as we have different error messages in our
version compared to the RISC OS Pyromaniac version.