Fixes from tests; defaulted tokens; fixed test buffers.
The test buffer and the message buffer were being allocated in the
same place. This meant that we overwrote the start of the buffer
with the results of lookups. This should now be fixed and we'll
get some better test output now.

Token defaults are now implemented and we should end up with the
right values coming out of the system for defaulted tokens.